Add ZFS compatibility
To use xfstests as a test suite for zfs it has been modified
slighty to handle zfs pooled storage.  The test suite is built
around the idea TEST/SCRATCH_DEV is a block device which backs
the filesystem.  That's not going to be true for zfs style
pooled storage.

To handle zfs pools when the filesystem type is set to 'zfs' then
TEST/SCRATCH_DEV is expected to be of the form '<pool>/<dataset>'.
The pool must be created prior to running 'check -zfs' but the
required datasets will be dynamically created.

You may choose to manually create the required dataset however
the 'mountpoint' property must be set to 'legacy'.  This allows
xfstests to use the standard Linux mount/umount tool chain to
mount the filesystems.  Otherwise it would need to use the
'zfs mount' command which would require more extensive changes
to xfstests.  Since the point of these tests is to validate the
VFS interfaces using the Linux mount command is fine.

Usage: ./check -zfs