Set refquota on when fssize provided
When a filesystem of a specific size is requested the the refquota.
This is going to be an artificial limit since zfs uses pooled storage
but that can be helpful when running tests which are designed to fill
the entire filesystem.

Additionally, update the scratch fs creation to first destroy the
previous dataset and then create a new one.