xfstests: Add support for zfs filesystems
Extend the existing xfstests infrastructure so it can be used to
test the zfs filesystems.

If blkid can property identify the device as a zfs member this
will get automatically set.  However, older version of blkid were
not ZFS-aware so it was helpful add the -zfs option to force zfs.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>