xfstests 015: Handle defered free space for ZFS
After removing a file in ZFS that space is not immediately visible.
It will become visible to tools such as df one of the next txg is
written and the pool state is durable.  To accomidate this remount
the filesystem to force a pool sync.  Alternately the test could
sleep for 5 seconds which is the default txg sync timeout.