Pickup upstream xfstest fix
Until we can refresh xfstests to the last version we need to
pull in the following upstream change to resolve false positives
for a few test cases.

See upstream commit 774f4dd775340adc53565dce858882de0d4e6e85
xfstests: unify apostrophes in output files

With coreutils v8.16 the style of apostrophes changed from `word' to
'word'. This is breaking some tests which use the older form.

This commit introduces function changes the golden output of the
affected tests and introduces a filter for the older style output.

[dchinner: modified to use a global filter in check rather than
per-test filters]
[rjohnston: minor comment change]

Signed-off-by: Tomas Racek <tracek@redhat.com>
Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Rich Johnston <rjohnston@sgi.com>